style improvements and favorites button
