+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@emotion/react": "11.9.0",
     "@emotion/styled": "11.8.1",
+    "@mui/icons-material": "5.8.2",
     "@mui/material": "5.8.2",
     "@reduxjs/toolkit": "1.8.2",
     "@testing-library/jest-dom": "5.16.4",
@@ -14,6 +15,7 @@
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-redux": "8.0.2",
+    "react-router-dom": "6.3.0",
     "react-scripts": "5.0.1",
     "web-vitals": "2.1.4"
   },
