+++ src/app/store.js
@@ -1,6 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
-import favoritesReducer from '../cityReducer';
+import favoritesReducer from '../favoritesReducer';
 
 export const store = configureStore({
-  reducer: favoritesReducer 
+  reducer: {
+    favoritesReducer
+  } 
 });
