+++ src/components/MainCityDisplay.jsx
@@ -2,15 +2,16 @@ import { Grid } from "@mui/material";
 import { Container, Gr } from "@mui/system";
 import React from "react";
 import DayForecast from "./dayForecast";
+import FavoriteButton from "./favoriteButton";
+import CityHeader from "./cityHeader";
 
 function CityDisplay({ weatherData, cityData }) {
     const dailyForeCasts = weatherData["DailyForecasts"];
-    const { cityName, country } = cityData;
 
     return (
-        <Grid container spacing={2}>
+        <Grid container spacing={2} >
             <Grid item container justifyContent="center">
-                <div>{cityName} | {country}</div>
+                <CityHeader cityData={cityData} />
             </Grid>
             <Grid item justifyContent="center" container spacing={5}>
                 {dailyForeCasts.map((day, key) => {
