+++ src/components/SearchBar.jsx
@@ -1,18 +1,30 @@
+import { Button, TextField } from "@mui/material";
+import { Box } from "@mui/system";
 import React, { useState } from "react";
+import SearchIcon from '@mui/icons-material/Search';
 import { citySearch, getCityWeather } from "../utils/apiRequests";
 
 function SearchBar() {
     const [input, setInput] = useState("");
 
     const handleClick = () => {
-        getCityWeather(input);
+        if (input !== "") {
+            getCityWeather(input);
+        }
+
     }
 
     return (
-        <div>
-            <input type="text" onChange={(e) => setInput(e.target.value)} placeholder="Search City" />
-            <button onClick={handleClick}>Search</button>
-        </div>
+        <Box justifyContent="center" display="flex" >
+            <TextField onChange={(e) => setInput(e.target.value)} 
+            variant="outlined" 
+            label="Search City" 
+            size="small"
+            />
+            <Button onClick={handleClick} variant="contained">
+                <SearchIcon/>
+            </Button>
+        </Box>
     )
 }
