+++ src/components/favoritesPage.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+function FavoritesPage({ }) {
+
+    return(
+        <>
+            Favorites
+        </>
+    )
+}
+
+export default FavoritesPage;
