+++ src/favoritesReducer.js
